benchmark/metalearned/requirements.txt
Bottleneck==1.3.1
FRB==1.1.4
dill==0.3.1.1
feather-format==0.4.0
fire==0.2.1
gitpython==3.0.5
h5py==2.10.0
keras==2.3.1
numpy==1.17.4
pandas==0.23.4
patool==1.12
pytest==5.3.0
python-dateutil==2.8.1
torch==1.2.0
requests==2.22.0
scipy==1.3.2
seaborn==0.9.0
scikit-learn==0.21.3
statsmodels==0.10.1
stldecompose==0.0.5
tables==3.6.1
tqdm==4.38.0
xlrd==1.2.0
